requirements.txt
lxml==3.2.3
pyquery==1.2.4
python-dateutil==2.1
spec==0.11.1
